JSON file added, initialized in template
In other comits, I forgot to upload the package.json, I uploaded it now. I forget that, I installed the sass package. JSON file created. Passed the data in to template section. I adjusted the backgrounds to the final color. Two function created. One is responsible for date convert. The other function is responsible for columns in bottom grid system.